@@ -1,7 +1,12 @@
 <script lang="ts">
-	export let href = '';
+	interface Props {
+		href?: string;
+		children?: import('svelte').Snippet;
+	}
+
+	let { href = '', children }: Props = $props();
 </script>
 
 <a {href} class="p-2 cursor-pointer hover:dark:bg-white hover:dark:text-black hover:bg-black hover:text-white">
-	<slot />
+	{@render children?.()}
 </a>
