@@ -1,7 +1,12 @@
 <script lang="ts">
-	export let href = '';
+	interface Props {
+		href?: string;
+		children?: import('svelte').Snippet;
+	}
+
+	let { href = '', children }: Props = $props();
 </script>
 
 <a {href} class="cursor-pointer border-b-2">
-	<slot />
+	{@render children?.()}
 </a>
